KIA_v1/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
ctransformers==0.2.27
dataclasses-json==0.6.4
deepdiff==6.7.1
diskcache==5.6.3
emoji==2.10.1
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.21.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-experimental==0.0.53
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.22
llama_cpp_python==0.2.55
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pillow==10.2.0
pinecone-client==3.1.0
protobuf==4.21.12
py-cpuinfo==9.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pywatchman==1.4.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
unstructured==0.12.5
unstructured-client==0.21.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
